Finished limit on read command
This has made it possible to request an amount of cheeps to be read, which will give the user more control.

If there is no argument it will print all cheeps. Could be set to some default value.

if there is invalid argument it will print all cheeps. Could be set to some default value.

if there is integer argument it will print the choosen amount or every cheeps has been displayed.